requirement.txt
Django==2.2.3
gunicorn==19.9.0
mysqlclient==1.4.2.post1
Pillow==6.1.0
requests==2.22.0
django-cors-headers==3.0.2
djangorestframework==3.10.1